rename pair classes
